requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
pipenv==2020.6.2
psycopg2==2.8.5
requests==2.23.0